+++ handlers/test.go
@@ -5,6 +5,7 @@ import (
 	"context"
 	"encoding/json"
 	"io/ioutil"
+	"mime/multipart"
 	"net/http"
 	"net/http/httptest"
 	"strings"
@@ -32,6 +33,8 @@ type ChannelHandleTestCase struct {
 	Status   int
 	Response string
 	Headers  map[string]string
+
+	MultipartFormFields map[string]string
 
 	Name        *string
 	Text        *string
@@ -127,7 +130,7 @@ func ensureTestServerUp(host string) {
 }
 
 // utility method to make a request to a handler URL
-func testHandlerRequest(tb testing.TB, s courier.Server, path string, headers map[string]string, data string, expectedStatus int, expectedBody *string, requestPrepFunc RequestPrepFunc) string {
+func testHandlerRequest(tb testing.TB, s courier.Server, path string, headers map[string]string, data string, multipartFormFields map[string]string, expectedStatus int, expectedBody *string, requestPrepFunc RequestPrepFunc) string {
 	var req *http.Request
 	var err error
 	url := fmt.Sprintf("https://%s%s", s.Config().Domain, path)
@@ -143,6 +146,21 @@ func testHandlerRequest(tb testing.TB, s courier.Server, path string, headers ma
 		} else if strings.Contains(data, "<") && strings.Contains(data, ">") {
 			contentType = "application/xml"
 		}
+		req.Header.Set("Content-Type", contentType)
+	} else if multipartFormFields != nil {
+		var body bytes.Buffer
+		bodyMultipartWriter := multipart.NewWriter(&body)
+		for k, v := range multipartFormFields {
+			fieldWriter, err := bodyMultipartWriter.CreateFormField(k)
+			require.Nil(tb, err)
+			_, err = fieldWriter.Write([]byte(v))
+			require.Nil(tb, err)
+		}
+		contentType := fmt.Sprintf("multipart/form-data;boundary=%v", bodyMultipartWriter.Boundary())
+		bodyMultipartWriter.Close()
+
+		req, err = http.NewRequest(http.MethodPost, url, bytes.NewReader(body.Bytes()))
+		require.Nil(tb, err)
 		req.Header.Set("Content-Type", contentType)
 	} else {
 		req, err = http.NewRequest(http.MethodGet, url, nil)
@@ -350,7 +368,7 @@ func RunChannelTestCases(t *testing.T, channels []courier.Channel, handler couri
 			mb.ClearQueueMsgs()
 			mb.ClearSeenExternalIDs()
 
-			testHandlerRequest(t, s, testCase.URL, testCase.Headers, testCase.Data, testCase.Status, &testCase.Response, testCase.PrepRequest)
+			testHandlerRequest(t, s, testCase.URL, testCase.Headers, testCase.Data, testCase.MultipartFormFields, testCase.Status, &testCase.Response, testCase.PrepRequest)
 
 			// pop our message off and test against it
 			contactName := mb.GetLastContactName()
@@ -433,14 +451,14 @@ func RunChannelTestCases(t *testing.T, channels []courier.Channel, handler couri
 		t.Run("Queue Error", func(t *testing.T) {
 			mb.SetErrorOnQueue(true)
 			defer mb.SetErrorOnQueue(false)
-			testHandlerRequest(t, s, validCase.URL, validCase.Headers, validCase.Data, 400, Sp("unable to queue message"), validCase.PrepRequest)
+			testHandlerRequest(t, s, validCase.URL, validCase.Headers, validCase.Data, validCase.MultipartFormFields, 400, Sp("unable to queue message"), validCase.PrepRequest)
 		})
 	}
 
 	if !validCase.NoInvalidChannelCheck {
 		t.Run("Receive With Invalid Channel", func(t *testing.T) {
 			mb.ClearChannels()
-			testHandlerRequest(t, s, validCase.URL, validCase.Headers, validCase.Data, 400, Sp("channel not found"), validCase.PrepRequest)
+			testHandlerRequest(t, s, validCase.URL, validCase.Headers, validCase.Data, validCase.MultipartFormFields, 400, Sp("channel not found"), validCase.PrepRequest)
 		})
 	}
 }
@@ -461,7 +479,7 @@ func RunChannelBenchmarks(b *testing.B, channels []courier.Channel, handler cour
 
 		b.Run(testCase.Label, func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				testHandlerRequest(b, s, testCase.URL, testCase.Headers, testCase.Data, testCase.Status, nil, testCase.PrepRequest)
+				testHandlerRequest(b, s, testCase.URL, testCase.Headers, testCase.Data, testCase.MultipartFormFields, testCase.Status, nil, testCase.PrepRequest)
 			}
 		})
 	}